feat: add LobbyScene + lobby ready/start flow; refactor LoginScene/startNetwork; remove DOM join overlay; fix duplicate handlers

+++ public/network.js
@@ -1,8 +1,8 @@
 let socket, myId, players = {}, playerColours = {};
 
-function connectToServer(onInit, onUpdate, onRemove) {
+function connectToServer(onInit, onUpdate, onRemove, onLobbyUpdate, onStartBattle, onGameStart) {
     socket = io();
-    
+
     // If a player name was entered before connecting, tell the server once connected
     socket.on('connect', () => {
         // expose socket to window so index overlay can use it
@@ -18,7 +18,7 @@ function connectToServer(onInit, onUpdate, onRemove) {
         for (let id in data.players) {
             playerColours[id] = data.players[id].colour; // server-assigned colour
         }
-        onInit(data.players);
+        if (onInit) onInit(data.players);
     });
 
     // New player joined or movement update
@@ -30,13 +30,28 @@ function connectToServer(onInit, onUpdate, onRemove) {
             playerColours[id] = position.colour;
         }
 
-        onUpdate(id, position);
+        if (onUpdate) onUpdate(id, position);
     });
 
     // Player disconnected
     socket.on('remove', id => {
         if (playerColours[id]) delete playerColours[id];
-        onRemove(id);
+        if (onRemove) onRemove(id);
+    });
+
+    // Lobby updates (list of players + ready flags)
+    socket.on('lobbyUpdate', data => {
+        if (onLobbyUpdate) onLobbyUpdate(data);
+    });
+
+    // Server says all players should start
+    socket.on('startBattle', () => {
+        if (onStartBattle) onStartBattle();
+    });
+
+    // Server sends full game start payload
+    socket.on('gameStart', data => {
+        if (onGameStart) onGameStart(data.players);
     });
 }
 
@@ -44,5 +59,17 @@ function connectToServer(onInit, onUpdate, onRemove) {
 function sendMove(position) {
     if (socket && socket.connected) {
         socket.emit('move', position);
+    }
+}
+
+function sendReady(isReady) {
+    if (socket && socket.connected) {
+        socket.emit('ready', !!isReady);
+    }
+}
+
+function requestStartBattle() {
+    if (socket && socket.connected) {
+        socket.emit('startBattle');
     }
 }
